Update MaterialRequisition status in an atomic transaction
This commit adds a new function, `update_material_requisition_status()`, which updates the status of a MaterialRequisition object to 'approved' using a Django serializer.

The function is decorated with the `@transaction.atomic` decorator to ensure that the database updates are performed in an atomic transaction. This means that if an error occurs during the update, the entire transaction will be rolled back and the database will be left in its original state.