feat: Update manufacturing order status in consumption transaction
This commit adds a new function, `update_manufacturing_order_status()`, which updates the status of a manufacturing order to 'mr_approved' using a Django serializer. The function is called within the `create_consumption_transaction()` function to update the manufacturing order status when a consumption transaction is created.

The `update_manufacturing_order_status()` function is decorated with the `@transaction.atomic` decorator to ensure that the database updates are performed in an atomic transaction. This guarantees data integrity and rollback of changes if an error occurs during the transaction.

The commit also includes necessary imports and updates to the `ManufacturingOrder` model to add the 'mr_approved' status choice.

```